Make Partition cache graph size
This one is provides an enormous speedup.
With some tracing and performance measurement, it turned out that
G.size(weight) was indeed very expensive, as this has been recomputed
from scratch by iterating over *all nodes* in the graph.

As this value never changes over the lifetime of a partition, the
partition will calculate this value one and make it available to the
quality metrics, which use it.