Fix non-termination of Leiden algorithm
This one took some effort to debug.
Turns out that this is a problem with the original algorithm, if the parameters chosen don't fit the graph.

Lemma 8 of the supplementary material to the "From Louvain to Leiden" paper already states that the algorithm will find an asymptotically stable partition.
However, the algorithm (as provided in the pseudocode) will not necessarily terminate, if the parameter γ to the Leiden algorithm is chosen too large.
(A more detailed description can be found in this changeset.)

This fix is rather simple:
The algorithm now remembers the previous partition and if it matches the newly calculated partition, it is clear that the sequence of partitions has become stationary and the algorithm can be terminated.